+++ notifications/views.py
@@ -6,7 +6,7 @@ from django.conf import settings
 from django.http import HttpResponseRedirect
 
 def notifications (request):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.is_verified:
         notifs = Notif.objects.filter(user=request.user).order_by('-timestamp') 
         context = {'notifs':notifs, 'user':request.user}
         return render(request, 'notifications/notifications.html', context)
@@ -15,7 +15,7 @@ def notifications (request):
         return redirect('dashboard')
     
 def delete_notification (request, pk):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.is_verified:
         notif = Notif.objects.get(user=request.user, pk = pk)
         notif.delete()
         return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
